cpu_support: Ask Linux for the number of cores instead of using cpuid
It turns out that getting it directly from the processor is *hard.*

+++ cpu_support.h
@@ -15,7 +15,7 @@ typedef struct {
 } cpu_support_cat_t;
 
 typedef struct {
-	uint8_t num_cores;
+	int num_cores;
 	bool hthreaded;
 	int num_cat_levels;
 	cpu_support_cat_t *cat_levels;
